[Backend] Moved controller from interfaces to http models
